+++ shiny/render/__init__.py
@@ -26,6 +26,8 @@ from ._express import (
 from ._render import (
     code,
     download,
+    download_button,
+    download_link,
     image,
     plot,
     table,
@@ -44,6 +46,8 @@ __all__ = (
     "table",
     "ui",
     "download",
+    "download_button",
+    "download_link",
     "DataGrid",
     "DataTable",
     "CellPatch",